refactor(wishlist): model item amount as UInt
Address PR #28 review: change the item `amount` field from `Int` to
`UInt` (defaults to `1u`, always >= 1). DB column stays INT — write via
`toInt()`, read via `toUInt()`. Edit view model parses input with
`toUIntOrNull().coerceAtLeast(1u)`; card/detail gate on `amount != 1u`.

+++ features/wishlist/common/src/jvmMain/kotlin/repo/ExposedWishlistItemRepo.kt
@@ -139,7 +139,7 @@ class ExposedWishlistItemRepo(
                 id = WishlistItemId(id),
                 wishlistId = WishlistId(get(wishlistIdColumn)),
                 title = get(titleColumn),
-                amount = get(amountColumn),
+                amount = get(amountColumn).toUInt(),
                 approximatePrice = amountOrNull(),
                 priceUnits = get(priceUnitsColumn),
                 links = linksFor(id),
@@ -169,7 +169,7 @@ class ExposedWishlistItemRepo(
     override fun update(id: WishlistItemId?, value: NewWishlistItem, it: UpdateBuilder<Int>) {
         it[wishlistIdColumn] = value.wishlistId.long
         it[titleColumn] = value.title
-        it[amountColumn] = value.amount
+        it[amountColumn] = value.amount.toInt()
         it[approxPriceIntColumn] = value.approximatePrice?.integerPart
         it[approxPriceDecColumn] = value.approximatePrice?.decimalPart?.toLong()
         it[priceUnitsColumn] = value.priceUnits
